Partial progress with translation classes
