+++ sdg/translations/__init__.py
@@ -1,4 +1,5 @@
 from .TranslationInputBase import TranslationInputBase
 from .TranslationInputSdmx import TranslationInputSdmx
+from .TranslationInputSdgTranslations import TranslationInputSdgTranslations
 from .TranslationOutputBase import TranslationOutputBase
 from .TranslationOutputJson import TranslationOutputJson
